member total subscribed days testing
