+++ src/DevBetterWeb.Core/ValueObjects/DateTimeRange.cs
@@ -25,6 +25,12 @@ namespace DevBetterWeb.Core.ValueObjects
       return (end - StartDate).Days;
     }
 
+    public int ToDaysToDate(DateTime endDate)
+    {
+      var end = endDate;
+      return (end - StartDate).Days;
+    }
+
     public bool Contains(DateTime date)
     {
       if(date >= StartDate && date <= EndDate)
