+++ tests/DevBetterWeb.Tests/Core/Entities/MemberTests/BookHelpers.cs
@@ -1,6 +1,4 @@
 ﻿using DevBetterWeb.Core.Entities;
-using DevBetterWeb.Core.ValueObjects;
-using System;
 
 namespace DevBetterWeb.Tests.Core.Entities.MemberTests
 {
@@ -24,25 +22,6 @@ namespace DevBetterWeb.Tests.Core.Entities.MemberTests
         Author = "Ilyana Smith",
         Details = "this is a book",
         PurchaseUrl = "https://buyabook.com"
-      };
-    }
-  }
-
-  public static class SubscriptionHelpers
-  {
-    public static Subscription GetDefaultTestSubscription()
-    {
-      return new Subscription
-      {
-        Dates = new DateTimeRange(new DateTime(2020, 3, 14), DateTime.Today.AddDays(20))
-      };
-    }
-
-    public static Subscription GetSubscriptionWithPastEndDate()
-    {
-      return new Subscription
-      {
-        Dates = new DateTimeRange(new DateTime(2020, 3, 14), DateTime.Today.AddDays(-20))
       };
     }
   }
